Formated Code. Added Label for friends list and users list. Both for the sever view and the client view.
