+++ SausageChat/Chat_Server_window.xaml
@@ -36,7 +36,7 @@
 
 
 
-    <ItemsControl ItemsSource="{Binding Messages}" x:Name ="Messages" Background="White" Margin="0,0,10,58"  >
+    <ItemsControl ItemsSource="{Binding Messages}" x:Name ="Messages" Background="#FF419EEA" Margin="0,0,10,58" BorderBrush="#FF008CFF" Foreground="White"  >
 
       <ItemsControl.ItemTemplate>
 
@@ -45,16 +45,16 @@
           <StackPanel Orientation ="Vertical">
 
             <TextBlock Text="{Binding}" ></TextBlock>
-            </StackPanel>
+          </StackPanel>
         </DataTemplate>
       </ItemsControl.ItemTemplate>
 
 
     </ItemsControl>
 
-        <StackPanel Margin="0,320,10,-34">
-            <TextBox x:Name="Server_message_button" Height="33" TextWrapping="Wrap" Controls:TextBoxHelper.Watermark="Enter Server Message"  VerticalAlignment="Top" SpellCheck.IsEnabled="True" Margin="0,0,0,11" TextChanged="Server_message_button_TextChanged" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="White"/>
-        <Button x:Name="Send_message_Button" Content="Send" Height="33" VerticalAlignment="Top" RenderTransformOrigin="0.542,1.415" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="#FFFCFDFF" Click="Send_message_Button_Click"/>
+    <StackPanel Margin="0,320,10,-34">
+      <TextBox x:Name="Server_message_button" Height="35" TextWrapping="Wrap" Controls:TextBoxHelper.Watermark="Enter Server Message"  VerticalAlignment="Top" SpellCheck.IsEnabled="True" Margin="0,0,0,11" TextChanged="Server_message_button_TextChanged" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="White"/>
+      <Button x:Name="Send_message_Button" Content="Send" Height="33" VerticalAlignment="Top" RenderTransformOrigin="0.542,1.415" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="#FFFCFDFF" Click="Send_message_Button_Click"/>
     </StackPanel>
 
 
@@ -77,9 +77,10 @@
       <Button x:Name="Start_server" Content="Start Server" Height="44" VerticalAlignment="Top" Margin="0,10,15,0" Click="Start_server_Click" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="#FFFCFDFF"/>
     </StackPanel>
     <!--The item source should be bound to the Users list in the VM, and it can bind to the name property-->
-    <ListBox ItemsSource="{Binding ConnectedUsers}" x:Name="User_List" Grid.Column="2" Margin="10,0,0,78" SelectionChanged="ListBox_SelectionChanged" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="#FFFCFDFF">
+    <ListBox ItemsSource="{Binding ConnectedUsers}" x:Name="User_List" Grid.Column="2" Margin="10,25,0,78" SelectionChanged="ListBox_SelectionChanged" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="#FFFCFDFF">
     </ListBox>
-    <TextBox Grid.Column="2" Height="197" Margin="10,305,0,-125" TextWrapping="Wrap" Text="TextBox" VerticalAlignment="Top" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="White"/>
+    <TextBox Grid.Column="2" Height="197" Margin="10,306,0,-124" TextWrapping="Wrap" Text="TextBox" VerticalAlignment="Top" Background="#FF419EEA" BorderBrush="#FF008CFF" Foreground="White"/>
+    <Label  HorizontalContentAlignment="Center" Content="User List" Grid.Column="2" HorizontalAlignment="Left" Height="25" Margin="15,0,0,0" VerticalAlignment="Top" Width="271" Foreground="White"/>
 
   </Grid>
 </Controls:MetroWindow>
