+++ SausageChat/Networking/SausageServer.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -24,12 +24,16 @@ namespace SausageChat.Networking
             get
             {
                 return Vm.ConnectedUsers;
+        //TODO:Error.  Please, if you can, fix this. 
             }
             set
             {
                 Vm.ConnectedUsers = value;
-            }
-        }
+        //TODO:Error.  Please, if you can, fix this. 
+
+
+      }
+    }
         public static List<IPAddress> Blacklisted { get; set; } = new List<IPAddress>();
         public static IPEndPoint LocalIp { get; set; } = new IPEndPoint(IPAddress.Any, PORT);
