+++ SausageChat/SausageChatServer.csproj
@@ -72,7 +72,7 @@
     <Compile Include="Networking\ServerCommandResult.cs" />
     <Compile Include="Networking\SausageConnection.cs" />
     <Compile Include="ViewModel.cs" />
-    <Page Include="MainWindow.xaml">
+    <Page Include="Chat_Server_window.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -80,8 +80,8 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="MainWindow.xaml.cs">
-      <DependentUpon>MainWindow.xaml</DependentUpon>
+    <Compile Include="Chat_Server_window.xaml.cs">
+      <DependentUpon>Chat_Server_window.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
   </ItemGroup>
