+++ SausageChatClient/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:SausageChatClient"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Client_view_window.xaml">
     <Application.Resources>
     <ResourceDictionary>
       <ResourceDictionary.MergedDictionaries>
